Add result

+++ app/src/main/java/com/example/racingapp/domain/entity/Track.kt
@@ -3,5 +3,5 @@ package com.example.racingapp.domain.entity
 data class Track(
     val title: String,
     val country: String,
-    val length: Double
+    val length: Int // In meters
 )
